requirements.txt
django
djangorestframework
google-generativeai
python-socketio
eventlet
djangorestframework-simplejwt
django-filter
drf-yasg
cryptography